Match OpenAI Go SDK Responses timestamps
Fix #274 by decoding Responses created_at as float64, matching the official OpenAI Go SDK. Leave stream indices and database timestamps unchanged.

Add JSON and SSE regression tests for decimal timestamps, including response.created and response.completed. Verify with go test ./llm/oai.

+++ llm/oai/oai_responses.go
@@ -180,7 +180,7 @@ type responsesToolGrammar struct {
 type responsesResponse struct {
 	ID        string                `json:"id"`
 	Object    string                `json:"object"` // "response"
-	CreatedAt int64                 `json:"created_at"`
+	CreatedAt float64               `json:"created_at"`
 	Status    string                `json:"status"` // "completed", "incomplete", etc.
 	Model     string                `json:"model"`
 	Output    []responsesOutputItem `json:"output"`
